Add reusable bottom shadow view
