+++ Sources/Recycling/ListViews/FavoriteFolders/Subviews/FavoriteFoldersSearchBar.swift
@@ -4,9 +4,7 @@
 
 import UIKit
 
-final class FavoriteFoldersSearchBar: UIView {
-    private static let maxShadowRadius: CGFloat = 3
-
+final class FavoriteFoldersSearchBar: BottomShadowView {
     weak var delegate: UISearchBarDelegate? {
         didSet {
             searchBar.delegate = delegate
@@ -39,45 +37,15 @@ final class FavoriteFoldersSearchBar: UIView {
         fatalError("init(coder:) has not been implemented")
     }
 
-    // MARK: - Overrides
-
-    override func layoutSubviews() {
-        super.layoutSubviews()
-        // Make shadow to be on bottom
-        let radius = FavoriteFoldersSearchBar.maxShadowRadius
-        let rect = CGRect(x: 0, y: bounds.maxY - radius, width: bounds.width, height: radius)
-        layer.shadowPath = UIBezierPath(rect: rect).cgPath
-    }
-
     // MARK: - Setup
 
     func configure(withPlaceholder placeholder: String?) {
         searchBar.placeholder = placeholder
     }
 
-    func updateShadow(using scrollView: UIScrollView) {
-        let contentFrame = CGRect(
-            x: -scrollView.contentOffset.x,
-            y: scrollView.frame.minY - scrollView.contentOffset.y - scrollView.contentInset.top,
-            width: scrollView.contentSize.width,
-            height: scrollView.contentSize.height + scrollView.contentInset.top
-        )
-
-        let intersection = contentFrame.intersection(frame)
-        layer.shadowRadius = min(intersection.height * 0.2, FavoriteFoldersSearchBar.maxShadowRadius)
-    }
-
     // MARK: - Setup
 
     private func setup() {
-        backgroundColor = .white
-
-        layer.masksToBounds = false
-        layer.shadowOpacity = 0.3
-        layer.shadowRadius = 0
-        layer.shadowOffset = .zero
-        layer.shadowColor = UIColor.black.cgColor
-
         addSubview(searchBar)
 
         NSLayoutConstraint.activate([
